requirement.txt
streamlit==1.28.0
mysql-connector-python==8.0.33
pandas==2.0.0
plotly==5.15.0
protobuf==3.20.0  # Add this to prevent version conflicts